Rename util library
